flask/to_to_list/requirements.txt
Flask-Smorest
Flask-SQLAlchemy
Flask-Migrate
marshmallow
python-dotenv
mysqlclient
pymysql
